AP-6060: Fix DBTreeBrowser for Oracle database
The "Database Browser" didn't work properly for the Oracle. A Metadata
discovery took a lot of time because the Oracle DB has couple "slow"
schemas like SYS and SYSTEM. The fix will use schema filter when Oracle
driver was detected and do not use filter in the opposite case.

AP-6060 (Add Oracle Connector)